for lopp adds

+++ prefix_pthreads.c
@@ -196,10 +196,10 @@ static void* pthreadCalc(void* rank){
 	
 	
 	long long my_rank = (long long)rank;
-	int my_start = my_rank * size / thread_count;
-	int my_end = (my_rank + 1.0) * size / thread_count;
+	long my_start = my_rank * size / thread_count;
+	long my_end = (my_rank + 1.0) * size / thread_count;
 	printf("\nrank in %llu:\n",my_rank);
-	int myS=my_end-my_start;
+	long myS=my_end-my_start;
     
     //creating a temp array for local prefix solution
 	int tempArr[myS];
